Reorganize initialization of the build
